add plus operator
