--- src/commonTest/kotlin/com/hoc081098/flowext/PlusTest.kt
@@ -1,94 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2021-2023 Petrus Nguyễn Thái Học
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.hoc081098.flowext
-
-import com.hoc081098.flowext.utils.BaseTest
-import com.hoc081098.flowext.utils.TestException
-import com.hoc081098.flowext.utils.test
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertIs
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.flowOf
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class PlusTest : BaseTest() {
-  @Test
-  fun plusTwoFlow() = runTest {
-    val flow1 = flowOf(1, 2, 3)
-    val flow2 = flowOf(4, 5, 6)
-
-    (flow1 + flow2)
-      .test(
-        listOf(
-          Event.Value(1),
-          Event.Value(2),
-          Event.Value(3),
-          Event.Value(4),
-          Event.Value(5),
-          Event.Value(6),
-          Event.Complete,
-        ),
-      )
-  }
-
-  @Test
-  fun plusTwoFlow2() = runTest {
-    (flowOf("a", 2, 3) + flowOf(4, 5, 6))
-      .test(
-        listOf(
-          Event.Value("a"),
-          Event.Value(2),
-          Event.Value(3),
-          Event.Value(4),
-          Event.Value(5),
-          Event.Value(6),
-          Event.Complete,
-        ),
-      )
-  }
-
-  @Test
-  fun testPlus_firstFailureUpstream() = runTest {
-    val flow = flowOf(1, 2, 3)
-    val failureFlow = kotlinx.coroutines.flow.flow<Nothing> { throw TestException("Crash!") }
-    val expectation: suspend (List<Event<Int>>) -> Unit = { events ->
-      val message = assertIs<TestException>(events.single().errorOrThrow()).message
-      assertEquals("Crash!", message)
-    }
-
-    (failureFlow + flow)
-      .test(null, expectation)
-
-    (failureFlow + flow + flow)
-      .test(null, expectation)
-
-    (failureFlow + flow + flow + flow)
-      .test(null, expectation)
-
-    (failureFlow + flow + flow + flow + flow)
-      .test(null, expectation)
-  }
-}
